added calculation by days,ageFactor,riskFactor, length of trip and return the prices

+++ backend/src/main/java/com/roman/Insurance/ageCategories/AgeCategoryDto.java
@@ -1,6 +1,5 @@
 package com.roman.Insurance.ageCategories;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,7 +8,7 @@ public record AgeCategoryDto(
         String name,
         Integer minAge,
         Integer maxAge,
-        BigDecimal priceFactor,
+        double priceFactor,
         LocalDateTime createdAt,
         LocalDateTime updatedAt
